requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.4
django-appconf==1.0.2
django-compressor==2.1.1
django-photologue==3.7
django-sortedm2m==1.5.0
djangorestframework==3.6.3
drf-nested-routers==0.90.0
ExifRead==2.1.2
gunicorn==19.7.1
numpy==1.13.3+mkl
olefile==0.44
Pillow==4.2.1
pytz==2017.2
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
static3==0.7.0
